wrong type
darcs-hash:20070228102937-2fc9d-dabf8f1bfe1d89413a5644114020f3926ebb471a.gz

+++ itkShapeLabelObject.h
@@ -43,7 +43,7 @@ class ITK_EXPORT PhysicalSizeLabelObjectAccessor
 {
 public:
   typedef TLabelObject LabelObjectType;
-  typedef unsigned long AttributeType;
+  typedef double AttributeType;
 
   inline const AttributeType operator()( const LabelObjectType * labelObject )
     {
